+++ sdk/messaging/azservicebus/receiver.go
@@ -43,6 +43,10 @@ const (
 	SubQueueTransfer SubQueue = 2
 )
 
+// defaultReceiverIdleTime is the amount of time before we consider a non-session-receiver
+// link idle, and will recycle it.
+const defaultReceiverIdleTime = 5 * time.Minute
+
 // Receiver receives messages using pull based functions (ReceiveMessages).
 type Receiver struct {
 	receiveMode ReceiveMode
@@ -58,8 +62,8 @@ type Receiver struct {
 	mu        sync.Mutex
 	receiving bool
 
-	defaultDrainTimeout      time.Duration
 	defaultTimeAfterFirstMsg time.Duration
+	idleTracker              *internal.LocalIdleTracker
 
 	cancelReleaser *atomic.Value
 }
@@ -114,12 +118,17 @@ func applyReceiverOptions(receiver *Receiver, entity *entity, options *ReceiverO
 }
 
 type newReceiverArgs struct {
-	ns                  internal.NamespaceWithNewAMQPLinks
-	entity              entity
 	cleanupOnClose      func()
+	entity              entity
 	getRecoveryKindFunc func(err error) internal.RecoveryKind
-	newLinkFn           func(ctx context.Context, session amqpwrap.AMQPSession) (internal.AMQPSenderCloser, internal.AMQPReceiverCloser, error)
-	retryOptions        RetryOptions
+	// idleTimeout is the amount of time before a link is considered idle.
+	// If idleTimeout < 0 then the idle timeout is disabled.
+	// If idleTimeout == 0 then the default idle timeout is used (defaultReceiverIdleTime)
+	// Default is to use defaultReceiverIdleTime.
+	idleTimeout  time.Duration
+	newLinkFn    func(ctx context.Context, session amqpwrap.AMQPSession) (internal.AMQPSenderCloser, internal.AMQPReceiverCloser, error)
+	ns           internal.NamespaceWithNewAMQPLinks
+	retryOptions RetryOptions
 }
 
 var emptyCancelFn = func() string {
@@ -131,11 +140,19 @@ func newReceiver(args newReceiverArgs, options *ReceiverOptions) (*Receiver, err
 		return nil, err
 	}
 
+	var idleTracker *internal.LocalIdleTracker
+
+	if args.idleTimeout > 0 {
+		idleTracker = &internal.LocalIdleTracker{MaxDuration: args.idleTimeout}
+	} else if args.idleTimeout == 0 {
+		idleTracker = &internal.LocalIdleTracker{MaxDuration: defaultReceiverIdleTime}
+	}
+
 	receiver := &Receiver{
 		lastPeekedSequenceNumber: 0,
 		cleanupOnClose:           args.cleanupOnClose,
-		defaultDrainTimeout:      time.Second,
 		defaultTimeAfterFirstMsg: 20 * time.Millisecond,
+		idleTracker:              idleTracker,
 		retryOptions:             args.retryOptions,
 		cancelReleaser:           &atomic.Value{},
 	}
@@ -204,8 +221,20 @@ func (r *Receiver) ReceiveMessages(ctx context.Context, maxMessages int, options
 		return nil, errors.New("receiver is already receiving messages. ReceiveMessages() cannot be called concurrently")
 	}
 
-	messages, err := r.receiveMessagesImpl(ctx, maxMessages, options)
-	return messages, internal.TransformError(err)
+	for {
+		messages, err := r.receiveMessagesImpl(ctx, maxMessages, options)
+
+		// when we have a client-side idle error it just means we went too long without
+		// any activity. The user didn't make this happen (nor did the service) so we treat it
+		// as just an internal "state reset" and try the receive again.
+		//
+		// When this does happen there aren't any messages or anything to return.
+		if internal.IsLocalIdleError(err) {
+			continue
+		}
+
+		return messages, internal.TransformError(err)
+	}
 }
 
 // ReceiveDeferredMessagesOptions contains optional parameters for the ReceiveDeferredMessages function.
@@ -409,6 +438,10 @@ func (r *Receiver) receiveMessagesImpl(ctx context.Context, maxMessages int, opt
 		log.Writef(EventReceiver, "Failure when receiving messages: %s", result.Error)
 	}
 
+	if internal.IsLocalIdleError(result.Error) {
+		return nil, result.Error
+	}
+
 	// If the user does get some messages we ignore 'error' and return only the messages.
 	//
 	// Doing otherwise would break the idiom that people are used to where people expected
@@ -419,7 +452,8 @@ func (r *Receiver) receiveMessagesImpl(ctx context.Context, maxMessages int, opt
 	// function on Receiver).will have the same issue and will return the relevant error
 	// at that time
 	if len(result.Messages) == 0 {
-		if internal.IsCancelError(result.Error) || rk == internal.RecoveryKindFatal {
+		if internal.IsCancelError(result.Error) ||
+			rk == internal.RecoveryKindFatal {
 			return nil, result.Error
 		}
 
@@ -515,12 +549,27 @@ type fetchMessagesResult struct {
 //
 // Note, if you want to only receive prefetched messages send the parentCtx in
 // pre-cancelled. This will cause us to only flush the prefetch buffer.
-func (r *Receiver) fetchMessages(parentCtx context.Context, receiver amqpwrap.AMQPReceiver, count int, timeAfterFirstMessage time.Duration) fetchMessagesResult {
+func (r *Receiver) fetchMessages(usersCtx context.Context, receiver amqpwrap.AMQPReceiver, count int, timeAfterFirstMessage time.Duration) fetchMessagesResult {
+	firstReceiveCtx := usersCtx
+
+	if r.idleTracker != nil {
+		idleCtx, cancelIdleCtx := r.idleTracker.NewContextWithDeadline(usersCtx)
+		defer cancelIdleCtx()
+
+		firstReceiveCtx = idleCtx
+	}
+
+	start := time.Now()
+
 	// The first receive is a bit special - we activate a short timer after this
 	// so the user doesn't end up in a situation where we're holding onto a bunch
 	// of messages but never return because they never cancelled and we never
 	// received all 'count' number of messages.
-	firstMsg, err := receiver.Receive(parentCtx)
+	firstMsg, err := receiver.Receive(firstReceiveCtx)
+
+	if r.idleTracker != nil {
+		err = r.idleTracker.Check(usersCtx, start, err)
+	}
 
 	if err != nil {
 		// drain the prefetch buffer - we're stopping because of a
@@ -542,7 +591,7 @@ func (r *Receiver) fetchMessages(parentCtx context.Context, receiver amqpwrap.AM
 
 	// after we get one message we will try to receive as much as we can
 	// during the `timeAfterFirstMessage` duration.
-	ctx, cancel := context.WithTimeout(parentCtx, timeAfterFirstMessage)
+	ctx, cancel := context.WithTimeout(usersCtx, timeAfterFirstMessage)
 	defer cancel()
 
 	var lastErr error
@@ -571,7 +620,7 @@ func (r *Receiver) fetchMessages(parentCtx context.Context, receiver amqpwrap.AM
 			//
 			// If we cancel: we want a nil error since there's no failure. In that case parentCtx.Err() is nil
 			// If they cancel: we want to forward on their cancellation error.
-			Error: parentCtx.Err(),
+			Error: usersCtx.Err(),
 		}
 	} else {
 		return fetchMessagesResult{