+++ sdk/messaging/azservicebus/session_receiver.go
@@ -66,12 +66,16 @@ func newSessionReceiver(ctx context.Context, args newSessionReceiverArgs, option
 	}
 
 	r, err := newReceiver(newReceiverArgs{
-		ns:                  args.ns,
-		entity:              args.entity,
 		cleanupOnClose:      args.cleanupOnClose,
-		newLinkFn:           sessionReceiver.newLink,
+		entity:              args.entity,
 		getRecoveryKindFunc: internal.GetRecoveryKindForSession,
-		retryOptions:        args.retryOptions,
+		// It's a bit more complicated to do an idle timeout for sessions
+		// and there's limited gain - the renewing the session lock provides
+		// a guarantee that the link is still alive.
+		idleTimeout:  -1,
+		newLinkFn:    sessionReceiver.newLink,
+		ns:           args.ns,
+		retryOptions: args.retryOptions,
 	}, options)
 
 	if err != nil {