+++ sdk/azcore/CHANGELOG.md
@@ -1,17 +1,32 @@
 # Release History
 
-## 1.2.0 (Unreleased)
+## 1.3.0 (Unreleased)
+
+### Features Added
+* Added `BearerTokenOptions.AuthorizationHandler` to enable extending `runtime.BearerTokenPolicy`
+  with custom authorization logic
+* Added `Client` types and matching constructors to the `azcore` and `arm` packages.  These represent a basic client for HTTP and ARM respectively.
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+* Updated `internal` module to latest version.
+
+## 1.2.0 (2022-11-04)
 
 ### Features Added
 * Added `ClientOptions.APIVersion` field, which overrides the default version a client
   requests of the service, if the client supports this (all ARM clients do).
-
-### Breaking Changes
+* Added package `tracing` that contains the building blocks for distributed tracing.
+* Added field `TracingProvider` to type `policy.ClientOptions` that will be used to set the per-client tracing implementation.
 
 ### Bugs Fixed
 * Fixed an issue in `runtime.SetMultipartFormData` to properly handle slices of `io.ReadSeekCloser`.
 * Fixed the MaxRetryDelay default to be 60s.
 * Failure to poll the state of an LRO will now return an `*azcore.ResponseError` for poller types that require this behavior.
+* Fixed a bug in `runtime.NewPipeline` that would cause pipeline-specified allowed headers and query parameters to be lost.
 
 ### Other Changes
 * Retain contents of read-only fields when sending requests.