+++ sdk/storage/azblob/CHANGELOG.md
@@ -4,9 +4,17 @@
 
 ### Features Added
 
+* Added BlobDeleteType to DeleteOptions to allow access to 'Permanent' DeleteType.
+* Added [Set Blob Expiry API](https://learn.microsoft.com/rest/api/storageservices/set-blob-expiry).
+* Added method `ServiceClient()` to the `azblob.Client` type, allowing access to the underlying service client.
+
 ### Breaking Changes
 
+* Corrected the name for `saoid` and `suoid` SAS parameters in `BlobSignatureValues` struct as per [this](https://learn.microsoft.com/rest/api/storageservices/create-user-delegation-sas#construct-a-user-delegation-sas)
+
 ### Bugs Fixed
+
+* Corrected signing of User Delegation SAS. Fixes [#19372](https://github.com/Azure/azure-sdk-for-go/issues/19372) and [#19454](https://github.com/Azure/azure-sdk-for-go/issues/19454)
 
 ### Other Changes